requirements.txt
azure.ai.anomalydetector == 3.0.0b6
azure.core == 1.26.4

